Less scripted and more unit tests
