--- scripted-tests/run/exception-handling/test
@@ -1,1 +1,0 @@
-> run
