--- scripted-tests/run/integer-long-highestonebit/build.sbt
@@ -1,3 +1,0 @@
-ScalaNativePlugin.projectSettings
-
-scalaVersion := "2.11.8"
