--- scripted-tests/run/integer-long-highestonebit/test
@@ -1,1 +1,0 @@
-> run
