--- scripted-tests/run/java-lang-float-double/project/scala-native.sbt
@@ -1,8 +1,0 @@
-{
-  val pluginVersion = System.getProperty("plugin.version")
-  if (pluginVersion == null)
-    throw new RuntimeException(
-      """|The system property 'plugin.version' is not defined.
-         |Specify this property using the scriptedLaunchOpts -D.""".stripMargin)
-  else addSbtPlugin("org.scala-native" % "sbt-scala-native" % pluginVersion)
-}
