--- scripted-tests/run/java-lang-float-double/test
@@ -1,1 +1,0 @@
-> run
