--- scripted-tests/run/java-lang-integer/test
@@ -1,1 +1,0 @@
-> run
