--- scripted-tests/run/java-util-random/test
@@ -1,1 +1,0 @@
-> run
