--- scripted-tests/run/string-from-to-ctring/test
@@ -1,1 +1,0 @@
-> run
