+++ unit-tests/src/main/scala/java/lang/LongSuite.scala
@@ -120,4 +120,20 @@ object LongSuite extends tests.Suite {
     assert(new Long(Long.MAX_VALUE) == new Long(Long.MAX_VALUE))
     assert(new Long(Long.MIN_VALUE) == new Long(Long.MIN_VALUE))
   }
+
+  test("highestOneBit") {
+    assert(Long.highestOneBit(1) == 1L)
+    assert(Long.highestOneBit(2) == 2L)
+    assert(Long.highestOneBit(3) == 2L)
+    assert(Long.highestOneBit(4) == 4L)
+    assert(Long.highestOneBit(5) == 4L)
+    assert(Long.highestOneBit(6) == 4L)
+    assert(Long.highestOneBit(7) == 4L)
+    assert(Long.highestOneBit(8) == 8L)
+    assert(Long.highestOneBit(9) == 8L)
+    assert(Long.highestOneBit(63) == 32L)
+    assert(Long.highestOneBit(64) == 64L)
+    assert(Long.highestOneBit(Int.MaxValue) == 1073741824)
+    assert(Long.highestOneBit(Int.MaxValue + 1L) == 2147483648L)
+  }
 }
